Adds a first version of the magic button, and a "manifold" post-processing step
