+++ src/op_REMESHERS.py
@@ -6,134 +6,11 @@ import time
 from mathutils import Vector
 import numpy as np
 
-class BaseRemesher(bpy.types.Operator):
-    bl_idname = "bakemyscan.empty_remesher"
-    bl_label  = "Empty remersher structure"
-
-    bl_options = {"REGISTER", "UNDO"}
-
-    #For executable remeshers
-    tmp        = tempfile.TemporaryDirectory()
-    executable = None
-    results    = []
-    keepMaterials = False
-
-    #For remeshers which need to duplicate the object
-    workonduplis = False
-
-
-    @classmethod
-    def poll(self, context):
-        if self.executable is not None:
-            if executable == "":
-                return 0
-        if len(context.selected_objects)!=1 or context.active_object is None:
-            return 0
-        for o in context.selected_objects:
-            if o.type != "MESH":
-                return 0
-        if context.mode!="OBJECT" and context.mode!="SCULPT":
-            return 0
-        return 1
-    def invoke(self, context, event):
-        return context.window_manager.invoke_props_dialog(self)
-
-    #To be overridden
-    def setexe(self, context):
-        pass
-    def export(self, context):
-        pass
-    def reimport(self, context):
-        pass
-    def remesh(self, context):
-        pass
-    def status(self, context):
-        if self.results[2] != 0:
-            self.report({"ERROR"}, "Remesh error, look in the console")
-            print("OUTPUT:\n%s\nERROR:\n%s\CODE:\n%s" % self.results)
-            return{"CANCELLED"}
-
-    def preprocess(self, context):
-        self.initialobject   = context.active_object
-        self.existingobjects = [o for o in bpy.data.objects]
-        bpy.ops.object.mode_set(mode='OBJECT')
-        bpy.ops.object.transform_apply(location=False, rotation=True, scale=True)
-        #Operators working on a duplicated object
-        if self.workonduplis:
-            bpy.ops.object.duplicate()
-            self.copiedobject = context.scene.objects.active
-            #Apply the modifiers
-            for m in self.copiedobject.modifiers:
-                bpy.ops.object.modifier_apply(modifier=m.name)
-    def postprocess(self, context):
-        #Check that there is only one new object
-        newObjects = [o for o in bpy.data.objects if o not in self.existingobjects]
-        if len(newObjects)==0:
-            self.report({'ERROR'}, '0 new objects')
-        elif len(newObjects)>1:
-            self.report({'ERROR'}, '0 new objects')
-        else:
-            #Get the new object
-            self.new = newObjects[0]
-            #Make it selected and active
-            bpy.ops.object.select_all(action='DESELECT')
-            self.new.select=True
-            bpy.context.scene.objects.active = self.new
-            #Remove edges marked as sharp, and delete the loose geometry
-            bpy.ops.object.editmode_toggle()
-            bpy.ops.mesh.mark_sharp(clear=True)
-            bpy.ops.mesh.select_all(action='SELECT')
-            bpy.ops.mesh.delete_loose()
-            bpy.ops.object.editmode_toggle()
-            #Shade smooth and rename it
-            #bpy.ops.object.shade_smooth()
-            #bpy.context.object.data.use_auto_smooth = False
-            #context.active_object.name = self.initialobject.name + "." + self.bl_label.lower().replace(" ","")
-            #Remove hypothetical material
-            if not self.keepMaterials:
-                while len(context.active_object.material_slots):
-                    context.active_object.active_material_index = 0
-                    bpy.ops.object.material_slot_remove()
-            #Hide the old object
-            #self.initialobject.hide = True
-    def execute(self, context):
-        #Set the executable path
-        self.setexe(context)
-        #Preprocess
-        self.preprocess(context)
-        #Export
-        if self.executable is not None:
-            self.exporttime = time.time()
-            self.export(context)
-            self.exporttime = time.time() - self.exporttime
-        #Remesh
-        self.remeshtime = time.time()
-        self.remesh(context)
-        self.remeshtime = time.time() - self.remeshtime
-        #Check the output
-        if self.executable is not None:
-            self.status(context)
-        #Import
-        if self.executable is not None:
-            self.importtime = time.time()
-            try:
-                self.reimport(context)
-            except:
-                self.report({"ERROR"}, "Import error, look in the console")
-                print("OUTPUT:\n%s\nERROR:\n%s\CODE:\n%s" % self.results)
-            self.importtime = time.time() - self.importtime
-        #Post-process
-        self.postprocess(context)
-        #Show the wireframe (debug)
-        #context.object.show_wire = True
-        #context.object.show_all_edges = True
-        #Report
-        self.report({'INFO'}, 'Remeshed to %d polygons' % len(context.active_object.data.polygons))
-        return{'FINISHED'}
+from . import op_REMESHERS_BASE as base
 
 # External software
 
-class Quadriflow(BaseRemesher):
+class Quadriflow(base.BaseRemesher):
     bl_idname = "bakemyscan.remesh_quadriflow"
     bl_label  = "Quadriflow"
 
@@ -180,7 +57,7 @@ class Quadriflow(BaseRemesher):
             satflip = self.satflip,
         )
 
-class Instant(BaseRemesher):
+class Instant(base.BaseRemesher):
     bl_idname = "bakemyscan.remesh_instant"
     bl_label  = "Instant Meshes"
 
@@ -280,7 +157,7 @@ class Instant(BaseRemesher):
                 r = self.r if self.r!="r0" else None
             )
 
-class Mmgs(BaseRemesher):
+class Mmgs(base.BaseRemesher):
     bl_idname = "bakemyscan.remesh_mmgs"
     bl_label  = "Mmgs"
 
@@ -360,7 +237,7 @@ class Mmgs(BaseRemesher):
     def status(self, context):
         pass
 
-class Meshlab(BaseRemesher):
+class Meshlab(base.BaseRemesher):
     bl_idname = "bakemyscan.remesh_meshlab"
     bl_label  = "Meshlab"
 
@@ -435,38 +312,46 @@ class Meshlab(BaseRemesher):
     def reimport(self, context):
         bpy.ops.import_scene.obj(filepath=os.path.join(self.tmp.name, "tmp.o.obj"))
 
-
 # Custom methods
 
-class Basic(BaseRemesher):
+class Basic(base.BaseRemesher):
     bl_idname = "bakemyscan.remesh_decimate"
     bl_label  = "Basic decimate"
     workonduplis = True
 
-    limit    = bpy.props.IntProperty(name="limit",    description="Target faces", default=1500, min=50, max=500000)
-    vertex_group = bpy.props.BoolProperty(name="vertex_group", description="Use vertex group", default=True)
+    limit    = bpy.props.IntProperty(description="Target faces", default=1500, min=50, max=500000)
+    vertex_group = bpy.props.BoolProperty(description="Use vertex group", default=True)
+    factor = bpy.props.FloatProperty(description="Weight factor", default=0.5, min=0., max=1.)
 
     def draw(self, context):
-        self.layout.prop(self, "limit", text="target triangles")
+        self.layout.prop(self, "limit", text="Number of faces")
+        self.layout.prop(self, "vertex_group", text="Use weights")
 
     def remesh(self, context):
         lr = self.copiedobject
         hr = self.initialobject
-        lr.modifiers.new("decimate", type='DECIMATE')
-        lr.modifiers["decimate"].ratio = float(self.limit/len(lr.data.polygons))
-        lr.modifiers["decimate"].use_collapse_triangulate = True
-        bpy.ops.object.modifier_apply(modifier="decimate")
+        bpy.ops.object.modifier_add(type='DECIMATE')
+        bpy.context.object.modifiers["Decimate"].ratio = float(self.limit/len(lr.data.polygons))
+        bpy.context.object.modifiers["Decimate"].use_collapse_triangulate = True
+        if len(bpy.context.object.vertex_groups)>0 and self.vertex_group:
+            bpy.context.object.modifiers["Decimate"].vertex_group = bpy.context.object.vertex_groups[0].name
+            bpy.context.object.modifiers["Decimate"].vertex_group_factor = self.factor
+            bpy.context.object.modifiers["Decimate"].invert_vertex_group = True
+        bpy.ops.object.modifier_apply(modifier="Decimate")
 
-class Quads(BaseRemesher):
+class Quads(base.BaseRemesher):
     bl_idname = "bakemyscan.remesh_quads"
     bl_label  = "Dirty quads"
     workonduplis = True
 
     nfaces = bpy.props.IntProperty(name="nfaces",   description="Decimate ratio",  default=1500, min=50, max=200000)
     smooth = bpy.props.IntProperty(  name="smooth", description="Smoothing steps", default=1, min=0, max=15)
+    vertex_group = bpy.props.BoolProperty(name="vertex_group", description="Use vertex group", default=True)
+    factor = bpy.props.FloatProperty(description="Weight factor", default=0.5, min=0., max=1.)
 
     def draw(self, context):
         self.layout.prop(self, "nfaces",  text="Number of quads (min)")
+        self.layout.prop(self, "smooth", text="Use weights")
         self.layout.prop(self, "smooth", text="Relaxation steps")
 
     def remesh(self, context):
@@ -479,9 +364,9 @@ class Quads(BaseRemesher):
         #Decimate it
         print("Decimating with ratio %.4f" % ratio)
         bpy.ops.object.modifier_add(type='DECIMATE')
-        if len(bpy.context.object.vertex_groups)>0:
+        if len(bpy.context.object.vertex_groups)>0 and self.vertex_group:
             bpy.context.object.modifiers["Decimate"].vertex_group = bpy.context.object.vertex_groups[0].name
-            bpy.context.object.modifiers["Decimate"].vertex_group_factor = 1
+            bpy.context.object.modifiers["Decimate"].vertex_group_factor = self.factor
             bpy.context.object.modifiers["Decimate"].invert_vertex_group = True
         bpy.context.object.modifiers["Decimate"].ratio = ratio
         bpy.ops.object.modifier_apply(apply_as='DATA', modifier="Decimate")
@@ -521,24 +406,27 @@ class Quads(BaseRemesher):
         bpy.ops.object.editmode_toggle()
         """
 
-
-class Iterative(BaseRemesher):
+class Iterative(base.BaseRemesher):
     bl_idname = "bakemyscan.remesh_iterative"
     bl_label  = "Iterative method"
     workonduplis = True
 
     limit    = bpy.props.IntProperty(name="limit",    description="Target faces", default=1500, min=50, max=500000)
-    manifold = bpy.props.BoolProperty(name="manifold", description="Make manifold", default=False)
-    vertex_group = bpy.props.BoolProperty(name="vertex_group", description="Use vertex group", default=True)
+    vertex_group = bpy.props.BoolProperty(name="vertex_group", description="Use vertex group", default=False)
+    factor = bpy.props.FloatProperty(description="Weight factor", default=0.5, min=0., max=1.)
 
     def draw(self, context):
         self.layout.prop(self, "limit", text="target triangles")
-        self.layout.prop(self, "manifold", text="make manifold")
+        self.layout.prop(self, "vertex_group", text="use weight paint")
 
     def do_one_iteration(self):
         print("  -- Planar decimation")
         bpy.ops.object.modifier_add(type='DECIMATE')
         bpy.context.object.modifiers["Decimate"].decimate_type = 'DISSOLVE'
+        if len(bpy.context.object.vertex_groups)>0 and self.vertex_group:
+            bpy.context.object.modifiers["Decimate"].vertex_group = bpy.context.object.vertex_groups[0].name
+            bpy.context.object.modifiers["Decimate"].vertex_group_factor = self.factor
+            bpy.context.object.modifiers["Decimate"].invert_vertex_group = True
         bpy.ops.object.modifier_apply(apply_as='DATA', modifier="Decimate")
         print("  -- Ensuring triangles only")
         bpy.ops.object.editmode_toggle()
@@ -551,10 +439,10 @@ class Iterative(BaseRemesher):
         bpy.ops.object.modifier_apply(apply_as='DATA', modifier="Smooth")
         print("  -- Initial decimation")
         bpy.ops.object.modifier_add(type='DECIMATE')
-        bpy.context.object.modifiers["Decimate"].ratio = 0.8
+        bpy.context.object.modifiers["Decimate"].ratio = 0.75
         if len(bpy.context.object.vertex_groups)>0 and self.vertex_group:
             bpy.context.object.modifiers["Decimate"].vertex_group = bpy.context.object.vertex_groups[0].name
-            bpy.context.object.modifiers["Decimate"].vertex_group_factor = 0.75
+            bpy.context.object.modifiers["Decimate"].vertex_group_factor = self.factor
             bpy.context.object.modifiers["Decimate"].invert_vertex_group = True
         bpy.ops.object.modifier_apply(apply_as='DATA', modifier="Decimate")
         print("  -- Shrinkwrap")
@@ -602,6 +490,10 @@ class Iterative(BaseRemesher):
         lr.modifiers.new("decimate", type='DECIMATE')
         lr.modifiers["decimate"].ratio = float(self.limit/len(lr.data.polygons))
         lr.modifiers["decimate"].use_collapse_triangulate = True
+        if len(bpy.context.object.vertex_groups)>0 and self.vertex_group:
+            lr.modifiers["decimate"].vertex_group = bpy.context.object.vertex_groups[0].name
+            lr.modifiers["decimate"].vertex_group_factor = self.factor
+            lr.modifiers["decimate"].invert_vertex_group = True
         bpy.ops.object.modifier_apply(modifier="decimate")
 
         #Remove hypothetical material
@@ -609,213 +501,16 @@ class Iterative(BaseRemesher):
             context.active_object.active_material_index = 0
             bpy.ops.object.material_slot_remove()
 
-        # 6 - Remove non manifold and doubles
-        print("-- Making manifold")
+        # 6 - Remove doubles
+        print("-- Removing degenerates")
         bpy.ops.object.editmode_toggle()
         bpy.ops.mesh.remove_doubles(threshold=0.0001)
         bpy.ops.mesh.dissolve_degenerate(threshold=0.0001)
         bpy.ops.mesh.vert_connect_nonplanar()
-        if self.manifold:
-            addon_utils.enable("object_print3d_utils")
-            bpy.ops.mesh.print3d_clean_non_manifold()
         bpy.ops.object.editmode_toggle()
-
-
-#Make a model symetrical on the x axis
-class Symmetry(BaseRemesher):
-    bl_idname = "bakemyscan.symetrize"
-    bl_label  = "Symmetry"
-
-    workonduplis  = True
-    keepMaterials = True
-
-    center = bpy.props.EnumProperty(
-        items= (
-            ('bbox','bbox','bbox'),
-            ('cursor','cursor','cursor'),
-            #Maybe more to come in the future depending on user feedbacks!
-        ),
-        description="center",
-        default="bbox"
-    )
-    axis = bpy.props.EnumProperty(
-        items= (
-            ('-X','-X','-X'),
-            ('+X','+X','+X'),
-            ('-Y','-Y','-Y'),
-            ('+Y','+Y','+Y'),
-            ('-Z','-Z','-Z'),
-            ('+Z','+Z','+Z')
-        ),
-        description="axis",
-        default="-X"
-    )
-
-    def remesh(self, context):
-        lr = self.copiedobject
-        hr = self.initialobject
-
-        #Get the symmetry center depending on the method (maybe apply obj.matrix_world?)
-        cursor = bpy.context.scene.cursor_location
-
-        center, dim = None, None
-        if self.center == "bbox":
-            localBb = 0.125 * sum((Vector(b) for b in lr.bound_box), Vector())
-            center  = lr.matrix_world * localBb
-            if "X" in self.axis:
-                dim = lr.dimensions[0]
-            elif "Y" in self.axis:
-                dim = lr.dimensions[1]
-            elif "Z" in self.axis:
-                dim = lr.dimensions[2]
-        elif self.center == "cursor":
-            center = cursor.copy()
-            #Get the maximum distance between 3D cursor and bbox points
-            dim = 0
-            corners = [lr.matrix_world * Vector(v) for v in lr.bound_box]
-            #Find the distance
-            for corner in corners:
-                #Get the corner projected on the desired axis
-                cornProj, cursProj = 0, 0
-                if "X" in self.axis:
-                    cornProj, cursProj = corner[0], cursor[0]
-                elif "Y" in self.axis:
-                    cornProj, cursProj = corner[1], cursor[1]
-                elif "Z" in self.axis:
-                    cornProj, cursProj = corner[2], cursor[2]
-                #Compute the distance
-                dist   = np.sqrt((cornProj - cursProj)**2)
-                if dist > dim:
-                    dim = dist
-
-        #Compute the cube translation so that its face is on the center
-        offset = center.copy()
-        if self.axis=="-X":
-            offset[0] = offset[0] + 5*dim/2
-        if self.axis=="+X":
-            offset[0] = offset[0] - 5*dim/2
-        if self.axis=="-Y":
-            offset[1] = offset[1] +5* dim/2
-        if self.axis=="+Y":
-            offset[1] = offset[1] -5* dim/2
-        if self.axis=="-Z":
-            offset[2] = offset[2] + 5*dim/2
-        if self.axis=="+Z":
-            offset[2] = offset[2] - 5*dim/2
-
-        bpy.ops.mesh.primitive_cube_add(radius=5*dim / 2 , view_align=False, enter_editmode=False, location=offset)
-        cube = context.active_object
-
-        #Make the original object active once again
-        bpy.ops.object.select_all(action='DESELECT')
-        context.scene.objects.active = lr
-        lr.select = True
-
-        #boolean cut
-        bpy.ops.object.modifier_add(type='BOOLEAN')
-        lr.modifiers["Boolean"].operation = 'DIFFERENCE'
-        lr.modifiers["Boolean"].object    = cube
-        bpy.ops.object.modifier_apply(apply_as='DATA', modifier="Boolean")
-
-        #Remove the cube
-        bpy.data.objects.remove(cube)
-
-        #Make the original object active once again
-        bpy.ops.object.select_all(action='DESELECT')
-        context.scene.objects.active = lr
-        lr.select             = True
-
-        #Add a mirror modifier
-        bpy.ops.object.modifier_add(type='MIRROR')
-        mod = bpy.context.object.modifiers["Mirror"]
-        mod.use_clip = True
-        #Set the correct axis
-        if "Y" in self.axis:
-            mod.use_x = False
-            mod.use_y = True
-        if "Z" in self.axis:
-            mod.use_x = False
-            mod.use_z = True
-        #Add an empty at the cursor or bbox center
-        if self.center == "cursor":
-            bpy.ops.object.empty_add(type='PLAIN_AXES', location=cursor)
-        elif self.center == "bbox":
-            bpy.ops.object.empty_add(type='PLAIN_AXES', location=center)
-        empty = context.active_object
-        mod.mirror_object = empty
-        #Make the original object active once again
-        bpy.ops.object.select_all(action='DESELECT')
-        context.scene.objects.active = lr
-        lr.select             = True
-        #Apply
-        bpy.ops.object.modifier_apply(apply_as='DATA', modifier="Mirror")
-        #Remove the empty
-        bpy.data.objects.remove(empty)
-
-        #Remove faces with too big a number of polygons, created because of the boolean
-        bpy.ops.object.editmode_toggle()
-        bpy.ops.mesh.select_all(action='DESELECT')
-        bpy.ops.mesh.select_face_by_sides(number=8, type='GREATER')
-        bpy.ops.mesh.delete(type='ONLY_FACE')
-        bpy.ops.mesh.select_all(action='SELECT')
-        bpy.ops.mesh.remove_doubles()
-        bpy.ops.object.editmode_toggle()
-
-        #Make the original object active once again
-        bpy.ops.object.select_all(action='DESELECT')
-        context.scene.objects.active = lr
-        lr.select              = True
-
-        return {"FINISHED"}
-
-#Relax the topology
-class Relax(BaseRemesher):
-    bl_idname = "bakemyscan.relax"
-    bl_label  = "Relaxation"
-
-    workonduplis  = True
-    keepMaterials = True
-
-    smooth = bpy.props.IntProperty(  name="smooth", description="Relaxation steps", default=2, min=0, max=150)
-
-    def draw(self, context):
-        self.layout.prop(self, "smooth", text="Relaxation steps")
-
-    def remesh(self, context):
-        lr = self.copiedobject
-        hr = self.initialobject
-
-        #Add a few shrinkwrapping / smoothing iterations to relax the surface
-        for i in range(self.smooth):
-            bpy.ops.object.modifier_add(type='SHRINKWRAP')
-            bpy.context.object.modifiers["Shrinkwrap"].target = hr
-            bpy.ops.object.modifier_apply(apply_as='DATA', modifier="Shrinkwrap")
-            bpy.ops.object.select_all(action='TOGGLE')
-            bpy.ops.object.modifier_add(type='SMOOTH')
-            bpy.ops.object.modifier_apply(apply_as='DATA', modifier="Smooth")
-
-        #With one last small smoothing step
-        if self.smooth > 0:
-            bpy.ops.object.modifier_add(type='SMOOTH')
-            bpy.ops.object.modifier_apply(apply_as='DATA', modifier="Smooth")
-
-        #Make the original object active once again
-        bpy.ops.object.select_all(action='DESELECT')
-        context.scene.objects.active = lr
-        lr.select              = True
-
-        #Hide the original object?
-        hr.hide = True
-        #bpy.data.objects.remove(hr)
-
-        return {"FINISHED"}
-
-
-
 
 
 def register() :
-    bpy.utils.register_class(BaseRemesher)
     bpy.utils.register_class(Quadriflow)
     bpy.utils.register_class(Mmgs)
     bpy.utils.register_class(Instant)
@@ -823,11 +518,8 @@ def register() :
     bpy.utils.register_class(Basic)
     bpy.utils.register_class(Iterative)
     bpy.utils.register_class(Quads)
-    bpy.utils.register_class(Symmetry)
-    bpy.utils.register_class(Relax)
 
 def unregister() :
-    bpy.utils.unregister_class(BaseRemesher)
     bpy.utils.unregister_class(Quadriflow)
     bpy.utils.unregister_class(Mmgs)
     bpy.utils.unregister_class(Instant)
@@ -835,5 +527,3 @@ def unregister() :
     bpy.utils.unregister_class(Basic)
     bpy.utils.unregister_class(Iterative)
     bpy.utils.unregister_class(Quads)
-    bpy.utils.unregister_class(Symmetry)
-    bpy.utils.unregister_class(Relax)
